refactor: add route groups for navigation control
