+++ frontend/src/app/[locale]/(navigation)/difficulty/[level]/page.tsx
