+++ frontend/src/app/[locale]/(navigation)/difficulty/page.tsx
