+++ frontend/src/app/[locale]/(navigation)/learning/layout.module.scss
