+++ frontend/src/app/[locale]/(navigation)/learning/page.tsx
