+++ frontend/src/app/[locale]/(navigation)/meanings/[category]/page.tsx
