+++ frontend/src/app/[locale]/(navigation)/meanings/page.tsx
