+++ frontend/src/app/[locale]/(navigation)/settings/Settings.module.scss
