+++ frontend/src/app/[locale]/(navigation)/settings/page.tsx
