+++ frontend/src/app/[locale]/(no-navigation)/layout.tsx
@@ -1,0 +1,5 @@
+const NoNavigationLayout = ({ children }: { children: React.ReactNode }) => {
+  return <>{children}</>;
+};
+
+export default NoNavigationLayout;
