+++ frontend/src/app/[locale]/(no-navigation)/login/page.tsx
@@ -1,0 +1,3 @@
+export default function LoginPage() {
+  return <div>LoginPage</div>;
+}
